Refactor test structure
Test locations were all wrong.

PLAT-1836
